Move CA rotation and CA manage out of ca package
