+++ pkg/server/ca/camanage/journal.go
@@ -1,4 +1,4 @@
-package ca
+package camanage
 
 import (
 	"crypto/x509"
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"github.com/spiffe/spire/pkg/common/diskutil"
+	"github.com/spiffe/spire/pkg/server/ca"
 	"github.com/spiffe/spire/proto/private/server/journal"
 	"github.com/spiffe/spire/proto/spire/common"
 	"github.com/zeebo/errs"
@@ -72,7 +73,7 @@ func (j *Journal) Entries() *JournalEntries {
 	return proto.Clone(j.entries).(*JournalEntries)
 }
 
-func (j *Journal) AppendX509CA(slotID string, issuedAt time.Time, x509CA *X509CA) error {
+func (j *Journal) AppendX509CA(slotID string, issuedAt time.Time, x509CA *ca.X509CA) error {
 	j.mu.Lock()
 	defer j.mu.Unlock()
 
@@ -100,7 +101,7 @@ func (j *Journal) AppendX509CA(slotID string, issuedAt time.Time, x509CA *X509CA
 	return nil
 }
 
-func (j *Journal) AppendJWTKey(slotID string, issuedAt time.Time, jwtKey *JWTKey) error {
+func (j *Journal) AppendJWTKey(slotID string, issuedAt time.Time, jwtKey *ca.JWTKey) error {
 	j.mu.Lock()
 	defer j.mu.Unlock()
