+++ pkg/server/endpoints/config.go
@@ -23,6 +23,7 @@ import (
 	"github.com/spiffe/spire/pkg/server/authpolicy"
 	bundle_client "github.com/spiffe/spire/pkg/server/bundle/client"
 	"github.com/spiffe/spire/pkg/server/ca"
+	"github.com/spiffe/spire/pkg/server/ca/camanage"
 	"github.com/spiffe/spire/pkg/server/cache/dscache"
 	"github.com/spiffe/spire/pkg/server/catalog"
 	"github.com/spiffe/spire/pkg/server/endpoints/bundle"
@@ -53,8 +54,8 @@ type Config struct {
 	// Bundle endpoint configuration
 	BundleEndpoint bundle.EndpointConfig
 
-	// CA Manager
-	Manager *ca.Manager
+	// JWTKey publisher
+	JWTKeyPublisher camanage.JwtKeyPublisher
 
 	// Makes policy decisions
 	AuthPolicyEngine *authpolicy.Engine
@@ -117,7 +118,7 @@ func (c *Config) maybeMakeBundleEndpointServer() Server {
 
 func (c *Config) makeAPIServers(entryFetcher api.AuthorizedEntryFetcher) APIServers {
 	ds := c.Catalog.GetDataStore()
-	upstreamPublisher := UpstreamPublisher(c.Manager)
+	upstreamPublisher := UpstreamPublisher(c.JWTKeyPublisher)
 
 	return APIServers{
 		AgentServer: agentv1.New(agentv1.Config{
