+++ pkg/server/endpoints/endpoints_test.go
@@ -106,7 +106,7 @@ func TestNew(t *testing.T) {
 		Catalog:          cat,
 		ServerCA:         serverCA,
 		BundleEndpoint:   bundle.EndpointConfig{Address: tcpAddr},
-		Manager:          manager,
+		JWTKeyPublisher:          manager,
 		Log:              log,
 		Metrics:          metrics,
 		RateLimit:        rateLimit,
@@ -173,7 +173,7 @@ func TestNewErrorCreatingAuthorizedEntryFetcher(t *testing.T) {
 		Catalog:          cat,
 		ServerCA:         serverCA,
 		BundleEndpoint:   bundle.EndpointConfig{Address: tcpAddr},
-		Manager:          manager,
+		JWTKeyPublisher:          manager,
 		Log:              log,
 		Metrics:          metrics,
 		RateLimit:        rateLimit,
