+++ pkg/server/endpoints/middleware.go
@@ -14,7 +14,7 @@ import (
 	"github.com/spiffe/spire/pkg/server/api/middleware"
 	"github.com/spiffe/spire/pkg/server/api/rpccontext"
 	"github.com/spiffe/spire/pkg/server/authpolicy"
-	"github.com/spiffe/spire/pkg/server/ca"
+	"github.com/spiffe/spire/pkg/server/ca/camanage"
 	"github.com/spiffe/spire/pkg/server/datastore"
 	"github.com/spiffe/spire/proto/spire/common"
 	"github.com/spiffe/spire/test/clock"
@@ -53,8 +53,8 @@ func EntryFetcher(ds datastore.DataStore) middleware.EntryFetcher {
 	})
 }
 
-func UpstreamPublisher(manager *ca.Manager) bundle.UpstreamPublisher {
-	return bundle.UpstreamPublisherFunc(manager.PublishJWTKey)
+func UpstreamPublisher(jwtKeyPublisher camanage.JwtKeyPublisher) bundle.UpstreamPublisher {
+	return bundle.UpstreamPublisherFunc(jwtKeyPublisher.PublishJWTKey)
 }
 
 func AgentAuthorizer(log logrus.FieldLogger, ds datastore.DataStore, clk clock.Clock) middleware.AgentAuthorizer {
